in_same_thread_without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_message
