+++ lumen_runtime/src/otp/erlang/cancel_timer_2/test.rs
@@ -13,7 +13,8 @@ use crate::process::SchedulerDependentAlloc;
 use crate::scheduler::{with_process, with_process_arc};
 use crate::test::{
     cancel_timer_message, different_timeout_message, has_message, receive_message, run, strategy,
-    timeout_after, timeout_after_half_and_wait, timeout_message, with_timer_in_different_thread,
+    timeout_after, timeout_after_half, timeout_after_half_and_wait, timeout_message,
+    with_timer_in_different_thread, with_timer_in_same_thread,
 };
 
 #[test]
