+++ lumen_runtime/src/otp/erlang/cancel_timer_2/test/with_reference_timer_reference/with_list_options/with_async_false/with_info_true/with_local_reference/with_timer/in_same_thread.rs
@@ -1,44 +1,11 @@
 use super::*;
 
-use crate::test::{timeout_after_half, with_timer_in_same_thread};
+use crate::test::with_timer_in_same_thread;
 
 #[test]
 #[ignore]
 fn without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_message() {
-    with_timer_in_same_thread(|milliseconds, message, timer_reference, process| {
-        timeout_after_half(milliseconds);
-
-        let timeout_message = timeout_message(timer_reference, message, process);
-
-        assert!(!has_message(process, timeout_message));
-
-        let first_result = native(process, timer_reference, options(process));
-
-        assert!(first_result.is_ok());
-
-        let milliseconds_remaining = first_result.unwrap();
-
-        assert!(milliseconds_remaining.is_integer());
-
-        assert!(process.integer(0).unwrap() < milliseconds_remaining);
-        assert!(milliseconds_remaining <= process.integer(milliseconds / 2).unwrap());
-
-        // again before timeout
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(false.into())
-        );
-
-        timeout_after_half(milliseconds);
-
-        assert!(!has_message(process, timeout_message));
-
-        // again after timeout
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(false.into())
-        );
-    })
+    super::in_same_thread_without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_message(options);
 }
 
 #[test]
